alterações

+++ front-end/lib/classes/carteira_medico_class.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 class MedicoCarteira {
-  int valorConsulta;
+  double valorConsulta;
   String chavePix;
   String paciente;
